requirements.txt
pillow>=10.0.0
pillow-heif>=0.16.0
